// packages/api/src/database/entities/fabric.entity.d.ts
import { FabricType } from '@virgio/shared';
export declare class Fabric {
    fabricCode: string;
    gsm?: number;
    weight?: number;
    composition: string;
    washCare: string;
    cost: number;
    fabricType: FabricType;
    leadTime: number;
    supplierInfo?: string;
    inventoryQty: number;
    lowStockThreshold: number;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=fabric.entity.d.ts.map