// packages/api/src/database/entities/trim.entity.d.ts
export declare class Trim {
    trimCode: string;
    material: string;
    nomenclature: string;
    sizeSpecs?: string;
    finish?: string;
    cost: number;
    treatmentRequirements?: string;
    leadTime: number;
    supplierInfo?: string;
    inventoryQty: number;
    shelfLife?: number;
    lowStockThreshold: number;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=trim.entity.d.ts.map